pyproject.toml
[project]
name = "custom-response-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "discord>=2.3.2",
    "json5>=0.12.1",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.1",
]
